Scintilla: Remove unused variables
(integrated from komodo (trunk) change 74203 by ericp)

git-svn-id: http://svn.openkomodo.com/repos/openkomodo/trunk@10933 b87d16e7-d29f-4ee1-b5ec-817d645d335f

+++ contrib/patches/scintilla/lexers/LexPerl.cxx
@@ -945,7 +945,7 @@ void ColourisePerlDoc(unsigned int startPos, int length, int , // initStyle
     
     int previousState = SCE_PL_DEFAULT; // Used to restore the original state after a comment.
 
-    if (startPos < 0) {
+    if ((int) startPos < 0) {
         startPos = 0;
         state = SCE_PL_DEFAULT;
     }
